+++ src/realm/object-store/c_api/realm.hpp
@@ -20,6 +20,7 @@
 
 #include <realm/object-store/c_api/util.hpp>
 #include <realm/object-store/binding_context.hpp>
+#include <realm/object-store/binding_callback_thread_observer.hpp>
 
 namespace realm::c_api {
 
@@ -61,6 +62,49 @@ private:
     CallbackRegistry<> m_realm_changed_callbacks;
     CallbackRegistryWithVersion<> m_realm_pending_refresh_callbacks;
     CallbackRegistry<const Schema&> m_schema_changed_callbacks;
+};
+
+class CBindingThreadObserver : public realm::BindingCallbackThreadObserver {
+public:
+    using ThreadCallback = util::UniqueFunction<void()>;
+    using ErrorCallback = util::UniqueFunction<void(const char*)>;
+
+    static CBindingThreadObserver& create()
+    {
+        static CBindingThreadObserver instance;
+        return instance;
+    }
+
+    void set(ThreadCallback&& on_create, ThreadCallback&& on_destroy, ErrorCallback&& on_error)
+    {
+        m_create_callback = std::move(on_create);
+        m_destroy_callback = std::move(on_destroy);
+        m_error_callback = std::move(on_error);
+    }
+
+    void did_create_thread() override
+    {
+        if (m_create_callback)
+            m_create_callback();
+    }
+
+    void will_destroy_thread() override
+    {
+        if (m_destroy_callback)
+            m_destroy_callback();
+    }
+
+    void handle_error(std::exception const& e) override
+    {
+        if (m_error_callback)
+            m_error_callback(e.what());
+    }
+
+private:
+    CBindingThreadObserver() = default;
+    ThreadCallback m_create_callback;
+    ThreadCallback m_destroy_callback;
+    ErrorCallback m_error_callback;
 };
 
 } // namespace realm::c_api